vit_classifier/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.20.0
mmengine>=0.7.0
transformers>=4.30.0
wandb>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
